Migrate the JSON literal report to proper types
This makes it more plausible for this format to be made available in a subcrate to allow other
libraries/tools to easily deserialize it. Also makes it more obvious the places where we're
being weirdly inconsistent or exposing random internals because I was lazy/hasty.

A weird side-effect of this is that the order of the fields got shuffled around. It seems
the json macro will sort fields by name, while proper derives use the decl order. The latter
seems preferred anyway, but unfortunately we take a huge obfuscation churn.

This also contains part of the introduction of the concept of --output-format=json-cli
because I realized I wanted to do this half way through adding that. WHOOPS.